requirements.txt
argh==0.26.2
numpy==1.11.0
protobuf==3.0.0b2
pygtp==0.2
sgf==0.5
six==1.10.0
tensorflow==0.9.0rc0
